add controller to upload message
