+++ src/index.js
@@ -1,12 +1,10 @@
 const express = require("express");
 const path = require("path");
-const fs = require("fs");
 const multer = require("multer");
 const user = require("./database/models/userSchema");
-const {uploadAudio, deleteFileInServer, uploadImage} = require('./utils/cloudinaryHandler')
+const { uploadAudio, uploadImage, deleteFileInServer } = require("./utils/cloudinaryHandler");
+const { addMessage } = require("./controllers/messageController");
 const upload = multer({ dest: "uploads/" });
-
-
 
 const saveUser = async () => {
   console.log("SAVE USER");
@@ -22,22 +20,18 @@ const saveUser = async () => {
   }
 };
 
-const messageRoute = async (req, res) => {  
+const messageRoute = async (req, res) => {
   try {
-    if(req.files){
+    if (req.files) {
       const { audio, photo } = req.files;
       const filePaths = [audio[0].path, photo[0].path];
-      console.log("FILEs", filePaths);      
-      const aud = await uploadAudio(audio[0].path, audio[0].originalname);
-      console.log(aud);
-      await uploadImage(photo[0].path, photo[0].originalname);
-      await deleteFileInServer(filePaths);
+      const { secure_url: audioUrl } = await uploadAudio(audio[0].path, audio[0].originalname);
+      const { secure_url: imgUrl } = await uploadImage(photo[0].path, photo[0].originalname);
+      console.log("AUDIO RESULT", audioUrl, imgUrl);
       // deleteFileInServer(audio[0].path)
-      // deleteFileInServer(photo[0].path)
-
     }
-    res.status(200).json({ msg: "Test" });
-  } catch (error) {    
+    res.status(200).json({ msg: "Files upload successful" });
+  } catch (error) {
     res.status(500);
   }
 };
@@ -47,10 +41,17 @@ const initializeApp = () => {
   app.use(express.json());
   app.get("/", async (req, res) => {
     // await saveUser();
-    return res.status(200).sendFile(path.join(__dirname, './sample-page.html'))
+    return res.status(200).sendFile(path.join(__dirname, "./sample-page.html"));
   });
 
-  app.post("/message", upload.fields([{name: "audio", maxCount: 1}, {name:"photo", maxCount: 1}]), messageRoute);
+  app.post(
+    "/message",
+    upload.fields([
+      { name: "audio", maxCount: 1 },
+      { name: "photo", maxCount: 1 },
+    ]),
+    addMessage
+  );
 
   return app;
 };
